@bsalamat 2018-08-29
I ran the scheduler anti-affinity benchmark with 5000 nodes scheduling 3000 pods with anti-affinity. This PR improves performance of the benchmark by over 11x. Also, scheduling 10,000 pods in a 1,000 node cluster shows 6x performance improvment. This is while our master branch includes [a relatively recent performance improvement of 20x](https://github.com/kubernetes/kubernetes/pull/62211) for the inter-pod affinity. So, this PR combined with #62211 improve inter-pod affinity/anti-affinity performance by over 100x!

ref/ #25319 #47318 

Before:
```
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingAntiAffinity/5000Nodes/0Pods-12         	    3000	 183630475 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	565.141s
```

After:
```
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingAntiAffinity/5000Nodes/0Pods-12         	    3000	  12235825 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	50.970s
```

Scheduling 10,000 pods in a 1,000 node cluster:

Before:
```
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingAntiAffinity/1000Nodes/0Pods-12         	   10000	  33543824 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	341.631s
```

After:
```
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingAntiAffinity/1000Nodes/0Pods-12         	   10000	   5338628 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	59.633s
```