@ahmad-diaa 2018-09-03
/test pull-kubernetes-bazel-test